refactor : name control item to checkbox item
